Move basic dependencies to library
